Move context file
